The ADC/DAC part of the board is done
Now I just need to figure out how to do the GPIO breakout, but I have a lot more flexibility here. Ideally, I'd like to keep the PMOD connector format, however I'm not sure if that is going to be possible. If I can't do that due to routing issues, than I can probably just stick a 50-something pin header on the the top side of the board at the left edge, on the other side of the LPC connector and call it a day.