Rev 1.1
Multiple changes are made, as per suggestions made by the redditors of r/PrintedCircuitBoards. Firstly, two 50 ohm resistors were added in series with the DAC output in order to comply with the minimum peak output current of the op-amps of 50mA. Secondly, two M3 mounting holes were added in order to have the possibility to add standoffs. Thirdly, a clearance issue regarding jumpers near the LPC connector on the ML602 board was fixed by changing the board geometry to include a cutout for these.

Some other things to look into are the use of some ferrite beads for supplies of the analog ICs.